move stellar object pgs to own folder
