+++ src/ts/playgrounds/playgroundRegistry.ts
@@ -25,20 +25,20 @@ import { createSierpinskiScene } from "./anomalies/sierpinski";
 import { createAsteroidFieldScene } from "./asteroidField";
 import { createAtmosphericScatteringScene } from "./atmosphericScattering";
 import { createAutomaticLandingScene } from "./automaticLanding";
-import { createBlackHoleScene } from "./blackHole";
 import { createCharacterDemoScene } from "./character";
 import { createDarkKnightScene } from "./darkKnight";
 import { createDebugAssetsScene } from "./debugAssets";
 import { createDefaultScene } from "./default";
 import { createFlightDemoScene } from "./flightDemo";
 import { createHyperspaceTunnelDemo } from "./hyperspaceTunnel";
-import { createNeutronStarScene } from "./neutronStar";
 import { createOrbitalDemoScene } from "./orbitalDemo";
 import { createRingsScene } from "./rings";
 import { createSpaceStationScene } from "./spaceStation";
 import { createSpaceStationUIScene } from "./spaceStationUI";
 import { createStarMapScene } from "./starMap";
 import { createStarSystemViewScene } from "./starSystemView";
+import { createBlackHoleScene } from "./stellarObjects/blackHole";
+import { createNeutronStarScene } from "./stellarObjects/neutronStar";
 import { createTutorialScene } from "./tutorial";
 import { createXrScene } from "./xr";
