+++ tests/e2e/stellarObjects/neutronStar.spec.ts
@@ -1,6 +1,6 @@
 import { test } from "@playwright/test";
 
-import { renderAndSnap } from "./utils/renderSnap";
+import { renderAndSnap } from "../utils/renderSnap";
 
 test("The neutron star playground renders correctly", async ({ page }) => {
     await renderAndSnap(page, {
