+++ tests/e2e/stellarObjects/neutronStar.spec.ts-snapshots/baseline-linux.png
